Ported several filetype tests to rspec
The test::unit tests were testing quite a bit of code that was not really
concerned with Puppet::Util::FileType.  There was not a one-to-one map to
rspec.  Instead I wrote several new tests and ported only the tests that 'fit
in'.  The remaining test::unit filetype tests will be removed once I determine
that the code is either 1) tested elsewhere or 2) new tests are created in a
proper location.